Updated the CDVBarcodeScanner.mm and README
In the .mm file, deprecated methods were replaced except for one (willAnimateRotationToInterfaceOrientation) which will be addressed later.

README was updated to include some more details for the iOS portion of the code. The previous fix was to fix an issue where the scanner hangs if it is closed by swiping the modal window down.